+++ NEWS.md
@@ -8,6 +8,7 @@ complete changelog, see the git history for each version via the version links.
 ### Removed
 - Removed support for Ruby versions older than 2.2
 - Removed support for Rails versions older than 4.2
+- Removed all deprecated code from Clearance 1.x
 
 [2.0.0]: https://github.com/thoughtbot/clearance/compare/v1.14.1...2.0